K8S_redis_part2 everything running
we're better, but the apollo server cannot connect to redis and the
replica is not working. The replica issue appears to be a port problem
that perhaps is because the master is using the port it is trying to
listen with in the same pod ?? need to figure this out.